+++ src/parser/block/paragraph/mention/cite.test.ts
@@ -23,17 +23,17 @@ describe('Unit: parser/block/paragraph/mention/cite', () => {
     });
 
     it('valid', () => {
-      assert.deepStrictEqual(inspect(parser('>>0')), [['<span class="cite">&gt;<a class="anchor" href="?res=0" data-depth="1">&gt;0</a></span>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>>0 ')), [['<span class="cite">&gt;<a class="anchor" href="?res=0" data-depth="1">&gt;0</a></span>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>>0\n')), [['<span class="cite">&gt;<a class="anchor" href="?res=0" data-depth="1">&gt;0</a></span>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>>>0')), [['<span class="cite">&gt;&gt;<a class="anchor" href="?res=0" data-depth="2">&gt;0</a></span>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>>0\n1')), [['<span class="cite">&gt;<a class="anchor" href="?res=0" data-depth="1">&gt;0</a></span>', '<br>'], '1']);
-      assert.deepStrictEqual(inspect(parser('>>0\n>1')), [['<span class="cite">&gt;<a class="anchor" href="?res=0" data-depth="1">&gt;0</a></span>', '<br>'], '>1']);
-      assert.deepStrictEqual(inspect(parser('>>0\n>>')), [['<span class="cite">&gt;<a class="anchor" href="?res=0" data-depth="1">&gt;0</a></span>', '<br>'], '>>']);
-      assert.deepStrictEqual(inspect(parser('>>0\n>>1')), [['<span class="cite">&gt;<a class="anchor" href="?res=0" data-depth="1">&gt;0</a></span>', '<br>', '<span class="cite">&gt;<a class="anchor" href="?res=1" data-depth="1">&gt;1</a></span>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>>0\n>>>1')), [['<span class="cite">&gt;<a class="anchor" href="?res=0" data-depth="1">&gt;0</a></span>', '<br>', '<span class="cite">&gt;&gt;<a class="anchor" href="?res=1" data-depth="2">&gt;1</a></span>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>>>0\n>>')), [['<span class="cite">&gt;&gt;<a class="anchor" href="?res=0" data-depth="2">&gt;0</a></span>', '<br>'], '>>']);
-      assert.deepStrictEqual(inspect(parser('>>>0\n>>1')), [['<span class="cite">&gt;&gt;<a class="anchor" href="?res=0" data-depth="2">&gt;0</a></span>', '<br>', '<span class="cite">&gt;<a class="anchor" href="?res=1" data-depth="1">&gt;1</a></span>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>>0')), [['<span class="cite">&gt;<a href="?res=0" class="anchor" data-depth="1">&gt;0</a></span>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>>0 ')), [['<span class="cite">&gt;<a href="?res=0" class="anchor" data-depth="1">&gt;0</a></span>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>>0\n')), [['<span class="cite">&gt;<a href="?res=0" class="anchor" data-depth="1">&gt;0</a></span>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>>>0')), [['<span class="cite">&gt;&gt;<a href="?res=0" class="anchor" data-depth="2">&gt;0</a></span>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>>0\n1')), [['<span class="cite">&gt;<a href="?res=0" class="anchor" data-depth="1">&gt;0</a></span>', '<br>'], '1']);
+      assert.deepStrictEqual(inspect(parser('>>0\n>1')), [['<span class="cite">&gt;<a href="?res=0" class="anchor" data-depth="1">&gt;0</a></span>', '<br>'], '>1']);
+      assert.deepStrictEqual(inspect(parser('>>0\n>>')), [['<span class="cite">&gt;<a href="?res=0" class="anchor" data-depth="1">&gt;0</a></span>', '<br>'], '>>']);
+      assert.deepStrictEqual(inspect(parser('>>0\n>>1')), [['<span class="cite">&gt;<a href="?res=0" class="anchor" data-depth="1">&gt;0</a></span>', '<br>', '<span class="cite">&gt;<a href="?res=1" class="anchor" data-depth="1">&gt;1</a></span>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>>0\n>>>1')), [['<span class="cite">&gt;<a href="?res=0" class="anchor" data-depth="1">&gt;0</a></span>', '<br>', '<span class="cite">&gt;&gt;<a href="?res=1" class="anchor" data-depth="2">&gt;1</a></span>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>>>0\n>>')), [['<span class="cite">&gt;&gt;<a href="?res=0" class="anchor" data-depth="2">&gt;0</a></span>', '<br>'], '>>']);
+      assert.deepStrictEqual(inspect(parser('>>>0\n>>1')), [['<span class="cite">&gt;&gt;<a href="?res=0" class="anchor" data-depth="2">&gt;0</a></span>', '<br>', '<span class="cite">&gt;<a href="?res=1" class="anchor" data-depth="1">&gt;1</a></span>', '<br>'], '']);
     });
 
   });