+++ src/parser/inline/annotation.test.ts
@@ -42,7 +42,7 @@ describe('Unit: parser/inline/annotation', () => {
 
     it('nest', () => {
       assert.deepStrictEqual(inspect(parser('((`a`))')), [['<sup class="annotation"><code data-src="`a`">a</code></sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('((@a))')), [['<sup class="annotation"><a class="account" href="/@a">@a</a></sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('((@a))')), [['<sup class="annotation"><a href="/@a" class="account">@a</a></sup>'], '']);
       assert.deepStrictEqual(inspect(parser('((http://host))')), [['<sup class="annotation"><a href="http://host" target="_blank">http://host</a></sup>'], '']);
       assert.deepStrictEqual(inspect(parser('((![]{a}))')), [['<sup class="annotation">!<a href="a">a</a></sup>'], '']);
       assert.deepStrictEqual(inspect(parser('((<a>))')), [['<sup class="annotation">&lt;a&gt;</sup>'], '']);