+++ src/parser/inline/autolink/hashref.ts
@@ -1,18 +1,16 @@
-import { undefined } from 'spica/global';
 import { AutolinkParser } from '../../inline';
-import { rewrite, creator, open } from '../../../combinator';
+import { union, rewrite, context, open, convert, fmap, lazy } from '../../../combinator';
+import { link } from '../link';
 import { str } from '../../source';
-import { html } from 'typed-dom';
+import { define } from 'typed-dom';
 
-export const hashref: AutolinkParser.HashrefParser = creator(rewrite(
+export const hashref: AutolinkParser.HashrefParser = lazy(() => fmap(rewrite(
   open('#', str(/^[0-9]{1,16}(?![0-9A-Za-z'_]|[^\x00-\x7F\s])/)),
-  (source, { host, url }) => [[
-    html('a',
-      {
-        class: 'hashref',
-        target: url && url.origin !== host?.origin
-          ? '_blank'
-          : undefined,
-      },
-      source)
-  ], '']));
+  context({ syntax: { inline: {
+    link: true,
+    autolink: false,
+  }}},
+  convert(
+    source => `[${source}]{ ${source.slice(1)} }`,
+    union([link])))),
+  ([el]) => [define(el, { class: 'hashref', href: null })]));