+++ src/parser/inline/autolink/url.ts
@@ -1,4 +1,3 @@
-import { undefined } from 'spica/global';
 import { AutolinkParser } from '../../inline';
 import { union, some, validate, focus, rewrite, context, creator, convert, surround, open, lazy } from '../../../combinator';
 import { link } from '../link';
@@ -10,9 +9,9 @@ export const url: AutolinkParser.UrlParser = lazy(() => validate(['http://', 'ht
   open(
     /^https?:\/\/(?=[\x21-\x7E])/,
     focus(/^[\x21-\x7E]+/, some(union([bracket, some(unescsource, closer)])))),
+  context({ syntax: { inline: { link: true } } },
   convert(
     url => `{ ${url} }`,
-    context({ syntax: { inline: { link: undefined } } },
     union([link]))))));
 
 const bracket: AutolinkParser.UrlParser.BracketParser = lazy(() => creator(union([